% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSclassifier.R
\name{MSclassifier}
\alias{MSclassifier}
\title{Median-Supplement methods}
\usage{
MSclassifier(X, testset = NULL, method = "MSRandomForest")
}
\arguments{
\item{X}{a data frame of values of attributes (e.g. gene expression levels) and 
classes (e.g. receptor status phenotypes in breast cancer).  Samples are in rows while attributes are in columns.  
The last column of X should have the classes for all instances in X (e.g. receptor status phenotypes of samples). 
This form the training set.}

\item{testset}{the set of new instances to be classified. The Default is NULL. When set to NULL, the function 
returns only the model. To classify new instances, specify the data frame of the new instances as the test set. 
It should have the form (and attributes) of X.}

\item{method}{specifies whether to determine a median-supplement Random
Forest or median-supplement Naive Bayes. 
"MSRandomForest" infers median-supplement Random Forest. 
"MSNaiveBayes" applies the median-supplement Naive Bayes.
The default is median-supplement Random Forest.}
}
\description{
MSclassifier implements a median-supplement approach to machine learning, supporting 
complete compliance efforts by never missing sensitive sub-datasets or allowing some 
sub-datasets to escape the classification process when balancing overall dataset as 
required in traditional classification models, for an automated and effective binary 
classification for optimal decision making. The median-supplement approaches to 
machine learning were first used to decipher hormone and HER2 receptor status phenotypes
in breast cancer.
}
\note{
These methods are applied to binary classification problems such as 
finding receptor status phenotypes in breast cancer. 
They require unbalanced training data set
involving only two labels.

They enhance capacities of Naive Bayes and Random Forest
as predictive classifiers to determine the classes or labels of
test sets or new instances from a training set with unequal numbers 
of members of classes. For example, it can be used for identifying receptor status phenotypes
of breast cancer patients from gene expression
data with unequal numbers of members of receptor status
phenotypes. In this case, the method determines the presence or absence of
receptor status phenotype. 

The methods proceed by introducing a
predetermined number of supplementary instances based on
the median of each attribute (feature) of the 
training set. They then apply a Naive Bayes or a Random Forest
method to infer models for predicting classes of test sets or new instances of interest.
For example, the method could predict the class, receptor 
status phenotypes, of a breast cancer patient given a large gene expression
training data.
}
\examples{
data("her2")
data("testset")
Model <- MSclassifier(X=her2, testset=NULL, method="MSRandomForest")
predictions <- predict(Model, newdata=testset)
head(predictions)
table(predictions, testset$her2_status)
Predictions <- MSclassifier(her2,testset=tetset, method="MSRandomForest")
# MSNaiveBayes
Model <- MSclassifier(her2, testset=NULL, method="MSNaiveBayes")
predictions <- predict(Model, newdata=testset)
Predictions <- MSclassifier(her2,testset=tetset, method="MSNaiveBayes")
}
\references{
Adabor ES, Acquaah-Mensah GK. 
Machine learning approaches to decipher hormone
and HER2 receptor status phenotypes in breast cancer.
Brief Bioinf., 2017.doi: 10.1093/bib/bbx138
}
\author{
Adabor ES (emmanuelsadabor@gimpa.edu.gh), Acquaah-Mensah GK (George.Acquaah-Mensah@mcphs.edu), 
Mazandu GK (kuzamunu@aims.ac.za)
}
